Move MemoryStore to common package
