+++ packages/common/src/main.ts
@@ -2,4 +2,5 @@ export type * from './types.js';
 
 export * from './convert.js';
 export * from './multicodec.js';
+export * from './stores.js';
 export * from './type-utils.js';
